refactor(common): extract wasm-agnostic types to ironclaw_common
Moves pure-data types out of src/tools/wasm/ into ironclaw_common so they
can be used without the wasm-sandbox feature:

- capabilities: Capabilities, HttpCapability, EndpointPattern, etc.
- capabilities_schema: CapabilitiesFile, AuthCapabilitySchema, etc.
- credential_mapping: CredentialMapping, CredentialLocation
- ext_error: WasmError
- limits: ResourceLimits, FuelConfig (data only)
- oauth_refresh: OAuthRefreshConfig
- rate_limit: ToolRateLimitConfig, ToolDiscoverySummary
- storage: StoredWasmTool, StoredCapabilities, ToolStatus, TrustLevel, etc.

src/tools/wasm/{capabilities,capabilities_schema,error}.rs are now thin
re-export shims. WasmResourceLimiter (wasmtime-dependent) stays in
src/tools/wasm/limits.rs. Postgres/libSQL WasmToolStore impls stay in
src/tools/wasm/storage.rs behind their respective feature flags.
