+++ crates/ironclaw_common/Cargo.toml
@@ -13,7 +13,12 @@ repository = "https://github.com/nearai/ironclaw"
 dist = false
 
 [dependencies]
+blake3 = "1"
+chrono = { version = "0.4", features = ["serde"] }
 chrono-tz = "0.10"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+thiserror = "2"
 tracing = "0.1"
+url = "2"
+uuid = { version = "1", features = ["v4", "v5", "serde"] }
