+++ crates/ironclaw_common/src/lib.rs
@@ -1,6 +1,14 @@
 //! Shared types and utilities for the IronClaw workspace.
 
+pub mod capabilities;
+pub mod capabilities_schema;
+pub mod credential_mapping;
 mod event;
+pub mod ext_error;
+pub mod limits;
+pub mod oauth_refresh;
+pub mod rate_limit;
+pub mod storage;
 mod timezone;
 mod util;
